+++ Source/OS9exec_core/events.h
@@ -68,12 +68,12 @@
 /* the routines */
 void init_events(void);
 
-os9err evLink(char *evName, ulong *evId);
-os9err evUnLnk(ulong evId);
-os9err evCreat(char *evName, int evValue, short wInc, short sInc, ulong *evId);
+os9err evLink(char *evName, uint32_t *evId);
+os9err evUnLnk(uint32_t evId);
+os9err evCreat(char *evName, int evValue, short wInc, short sInc, uint32_t *evId);
 os9err evDelet(char *evName);
-os9err evWait(ulong evId, int minV, int maxV, int *evValue);
-os9err evSignl(ulong evId);
-os9err evSet(ulong evId, int evValue, int *prvValue);
+os9err evWait(uint32_t evId, int minV, int maxV, int *evValue);
+os9err evSignl(uint32_t evId);
+os9err evSet(uint32_t evId, int evValue, int *prvValue);
 
 /* eof */
